requirements.txt
matplotlib==3.2.1
numpy==1.18.4
openvr==1.11.1101
Pillow==7.1.2
scipy==1.4.1
tqdm==4.46.0
sqlite3